Feat : 사용자 확인 후 api 진행 및 레슨 생성자 아이디 생성 과정 변경

+++ src/main/java/com/example/nanuri/service/lesson/LessonService.java
@@ -7,10 +7,13 @@ import com.example.nanuri.domain.lesson.lessonImg.LessonImgId;
 import com.example.nanuri.domain.lesson.lessonImg.LessonImgRepository;
 import com.example.nanuri.dto.lesson.LessonRequestDto;
 import com.example.nanuri.dto.lesson.LessonResponseDto;
+import com.example.nanuri.handler.exception.AuthenticationForbiddenException;
+import com.example.nanuri.handler.exception.AuthenticationNullPointerException;
 import com.example.nanuri.handler.exception.ErrorCode;
 import com.example.nanuri.handler.exception.LessonNotFoundException;
 import com.example.nanuri.service.aws.S3Service;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -27,10 +30,14 @@ public class LessonService {
     private final S3Service s3Service;
 
     @Transactional
-    public void save(LessonRequestDto lessonRequestDto) {
-        Lesson lesson  = lessonRepository.save(lessonRequestDto.toEntity());
+    public void save(LessonRequestDto lessonRequestDto , Authentication authentication) {
+        Lesson lesson  = lessonRepository.save(lessonRequestDto.toEntity(Long.parseLong(authentication.getName())));
         if(lessonRequestDto.getImages()!=null) {
+            System.out.println(lessonRequestDto.getImages());
+            System.out.println(lessonRequestDto.getImages().isEmpty());
+            System.out.println(lessonRequestDto.getImages().size());
             for (MultipartFile multipartFile : lessonRequestDto.getImages()) {
+                System.out.println(multipartFile);
                 String lessonImg = s3Service.upload(multipartFile, "lessonImg");
                 lessonImgRepository.save(
                         LessonImg.builder()
@@ -59,30 +66,61 @@ public class LessonService {
     }
 
     @Transactional(readOnly = true)
-    public LessonResponseDto findById(int lessonId){
+    public LessonResponseDto findById(Long lessonId){
         Lesson lesson = lessonRepository.findById(lessonId)
                 .orElseThrow(()-> new LessonNotFoundException(ErrorCode.LESSON_NOT_FOUND));
         return new LessonResponseDto(lesson);
     }
 
     @Transactional
-    public void delete(int lessonId){
+    public void delete(Long lessonId , Authentication authentication){
+
+        // lesson 찾기
         Lesson lesson = lessonRepository.findById(lessonId)
                 .orElseThrow(() -> new LessonNotFoundException(ErrorCode.LESSON_NOT_FOUND));
+
+        // lesson의 생성자가 api 요청자와 동일한지 확인
+        if(!isAuthorizedUser(lesson.getCreator(),authentication)){
+            return;
+        }
+
+        // lesson 이미지 찾기
         List<LessonImg> lessonImgs = lessonImgRepository.findByLessonId(lessonId);
+
+        //S3에서 이미지 파일 삭제
         for(LessonImg lessonImg : lessonImgs){
             s3Service.deleteImage(lessonImg.getLessonImgId().getLessonImg());
         }
+
+        //DB 에 삭제
         lessonImgRepository.deleteAllByLessonId(lessonId);
         lessonRepository.delete(lesson);
 
     }
 
     @Transactional
-    public void updateStatus(int lessonId){
+    public void updateStatus(Long lessonId, Authentication authentication){
+
+        // lesson 찾기
         Lesson lesson = lessonRepository.findById(lessonId)
                 .orElseThrow(()-> new LessonNotFoundException(ErrorCode.LESSON_NOT_FOUND));
+
+        // lesson의 생성자가 api 요청자와 동일한지 확인
+        if(!isAuthorizedUser(lesson.getCreator(),authentication)){
+            return;
+        }
+
         lesson.updateStatus();
+    }
+
+    private boolean isAuthorizedUser(Long creatorId ,Authentication authentication){
+        if(authentication==null){
+            throw new AuthenticationNullPointerException(ErrorCode.NULL_AUTHENTICATION);
+        }
+        if( Long.parseLong(authentication.getName()) != creatorId){
+            throw new AuthenticationForbiddenException(ErrorCode.FORBIDDEN_AUTHENTICATION);
+        }
+        return true;
     }
 
 }
